requirements.txt
streamlit
langchain
langchain-community
langchain-pinecone
pinecone-client
ollama
python-dotenv